colocate variant methods
